Refactor to use new 'acl grant' feature from cli
