+++ package.json
@@ -23,9 +23,9 @@
     "coveralls": "cat ./coverage/lcov.info | npx coveralls",
     "ganache-cli:test": "sh ./node_modules/@aragon/test-helpers/ganache-cli.sh",
     "test": "cross-env TRUFFLE_TEST=true npm run ganache-cli:test",
-    "set-oracle": "npx truffle exec ./scripts/setTokenOracle.js",
-    "set-oracle:rinkeby": "npx truffle exec ./scripts/setTokenOracle.js --network rinkeby",
-    "set-oracle:mainnet": "npx truffle exec ./scripts/setTokenOracle.js --network mainnet"
+    "set-oracle": "./scripts/setTokenOracle.sh development",
+    "set-oracle:rinkeby": "./scripts/setTokenOracle.sh rinkeby",
+    "set-oracle:mainnet": "./scripts/setTokenOracle.sh mainnet"
   },
   "files": [
     "/arapp.json",
